Bussiness/nfce-scraper/requirements.txt
streamlit
opencv-python
pyzbar
selenium
beautifulsoup4
pandas
Pillow